eth, miner: move the public miner api into eth to access etherbase
